% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{add_contributor_guidelines}
\alias{add_contributor_guidelines}
\title{Add GSM Contributor Guidelines markdown to package}
\usage{
add_contributor_guidelines(strPackageDir = ".", overwrite = TRUE)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{overwrite}{(\code{boolean}) Overwrite existing files?}
}
\description{
Add GSM Contributor Guidelines markdown to package
}
