% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{add_gsm_actions}
\alias{add_gsm_actions}
\alias{add_gilead_actions}
\title{Add Gilead GitHub Actions to package}
\usage{
add_gsm_actions(strPackageDir = ".", overwrite = TRUE, verbose = TRUE)

add_gilead_actions(strPackageDir = ".", overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{strPackageDir}{String. Path to package directory}

\item{overwrite}{Logical. Overwrite existing files?}

\item{verbose}{Logical. Inform about changes?}
}
\value{
A character vector of added and updated action names, invisibly.
}
\description{
Add the official Gilead GitHub Actions from
\url{https://github.com/Gilead-BioStats/gsm.utils@actions-v1} to a package, and
update existing Gilead GitHub Actions to the latest versions if necessary.
}
