% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{add_pkgdown_examples_to_yaml}
\alias{add_pkgdown_examples_to_yaml}
\title{Add example menu items to pkgdown YAML}
\usage{
add_pkgdown_examples_to_yaml(pkgdown_yaml, html_files, metadata)
}
\arguments{
\item{pkgdown_yaml}{List representation of \verb{_pkgdown.yml}.}

\item{html_files}{Character vector of example HTML files.}

\item{metadata}{Data frame of example metadata.}
}
\value{
Updated pkgdown YAML list.
}
\description{
Add example menu items to pkgdown YAML
}
\keyword{internal}
