% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{add_pkgdown_nav}
\alias{add_pkgdown_nav}
\title{Add rendered assets to pkgdown navbar menu}
\usage{
add_pkgdown_nav(pkgdown_yml, menu_subdir, rendered_assets, verbose)
}
\arguments{
\item{pkgdown_yml}{Character. Path to the \verb{_pkgdown.yml} file to update with
new menu items, or from which to remove unused menu items.}

\item{menu_subdir}{Character. Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{rendered_assets}{Character. Path to successfully rendered HTML files to
add to pkgdown menu.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\value{
\code{NULL} (invisibly). Updates \verb{_pkgdown.yml} in place.
}
\description{
Add rendered assets to pkgdown navbar menu
}
\keyword{internal}
