% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{build_examples_menu}
\alias{build_examples_menu}
\title{Build ordered example menu}
\usage{
build_examples_menu(html_files, metadata)
}
\arguments{
\item{html_files}{Character vector of example HTML files.}

\item{metadata}{Data frame with \code{html}, \code{title}, and \code{index} columns.}
}
\value{
List of menu entries with \code{html} and \code{title}.
}
\description{
Build ordered example menu
}
\keyword{internal}
