% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slides.R
\name{build_slides_template}
\alias{build_slides_template}
\title{Build template content for a slide deck}
\usage{
build_slides_template(strTitle, intIndex)
}
\arguments{
\item{strTitle}{(\code{string}) Title for the slide deck.}

\item{intIndex}{(\code{numeric}) Optional ordering index for the slides menu.}
}
\value{
Character vector of template lines.
}
\description{
Build template content for a slide deck
}
\keyword{internal}
