% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{check_workflow_compliance}
\alias{check_workflow_compliance}
\title{Check Workflow Template Compliance}
\usage{
check_workflow_compliance(...)
}
\arguments{
\item{...}{Placeholder to prevent other errors.}
}
\value{
An error with instructions for updating.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function has been replaced by a GitHub Action to handle versions more
cleanly and automatically. Use \code{add_action("workflow-template-check.yaml")}
(or, more generally, \code{\link[=update_gsm_package]{update_gsm_package()}}) to replace this function.
}
