% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-check_workflow_compliance.R
\name{check_workflow_headers}
\alias{check_workflow_headers}
\title{Check Workflow File Headers}
\usage{
check_workflow_headers(
  workflows_dir,
  existing_workflows,
  expected_workflows,
  gsm_utils_version
)
}
\arguments{
\item{workflows_dir}{\code{character} path to workflows directory}

\item{existing_workflows}{\code{character} vector of existing workflow file names}

\item{expected_workflows}{\code{character} vector of expected workflow file names}

\item{gsm_utils_version}{\code{character} expected gsm.utils version}
}
\value{
\code{character} vector of header issues
}
\description{
Internal function to check version and generated-by headers in workflow files.
}
\keyword{internal}
