% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-construct_menu.R
\name{construct_md_metadata_table}
\alias{construct_md_metadata_table}
\title{Construct a data frame of menu metadata from a single md file}
\usage{
construct_md_metadata_table(md_file)
}
\value{
A data.frame with columns \code{html_file}, \code{title}, and \code{index} for one
md file.
}
\description{
Construct a data frame of menu metadata from a single md file
}
\keyword{internal}
