% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-construct_menu.R
\name{construct_menu}
\alias{construct_menu}
\title{Construct a list of menu entries for rendered assets}
\usage{
construct_menu(rendered_assets, metadata, existing_menu = NULL)
}
\arguments{
\item{rendered_assets}{Character. Path to successfully rendered HTML files to
add to pkgdown menu.}

\item{existing_menu}{List. Existing menu items from the pkgdown YAML
contents, if any, to preserve when adding new menu items for rendered
assets.}
}
\value{
A list of lists with \code{text} and \code{href} entries for each rendered
asset, sorted by \code{index} and \code{title} metadata from the corresponding md
file.
}
\description{
Construct a list of menu entries for rendered assets
}
\keyword{internal}
