% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-construct_menu.R
\name{construct_menu_metadata_table}
\alias{construct_menu_metadata_table}
\title{Construct a data frame of menu metadata from md files}
\usage{
construct_menu_metadata_table(menu_subdir)
}
\arguments{
\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}
}
\value{
A data.frame with columns \code{html_file}, \code{title}, and \code{index} for each
md file in the menu subdirectory.
}
\description{
Construct a data frame of menu metadata from md files
}
\keyword{internal}
