% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{.find_issue_path}
\alias{.find_issue_path}
\title{A simple path constructor for mocking}
\usage{
.find_issue_path(strPackageDir)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}
}
\value{
The path to issue templates.
}
\description{
A simple path constructor for mocking
}
\keyword{internal}
