% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asset.R
\name{.make_title_line}
\alias{.make_title_line}
\title{Build a title line for YAML front matter}
\usage{
.make_title_line(strTitle)
}
\arguments{
\item{strTitle}{(\code{string}) Display title.}
}
\value{
A single string, e.g. \code{'title: "My Title"'}.
}
\description{
Build a title line for YAML front matter
}
\keyword{internal}
