% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-shared_params.R
\name{.shared-params}
\alias{.shared-params}
\title{Shared parameters}
\arguments{
\item{assets_dir}{(\code{string}) Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}

\item{existing_menu}{List. Existing menu items from the pkgdown YAML
contents, if any, to preserve when adding new menu items for rendered
assets.}

\item{menu}{(\code{string}) Menu folder name.}

\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{output_dir}{(\code{string}) Directory to write the example to.}

\item{output_file}{(\code{string}) Path to output HTML file to create.}

\item{overwrite}{(\code{boolean}) Overwrite existing files? Default is \code{TRUE}.}

\item{params}{List. Optional list of parameters to pass to
\code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{pkgdown_yml}{(\code{string}) Path to the \verb{_pkgdown.yml} file to update with
new menu items, or from which to remove unused menu items.}

\item{qmd_file}{(\code{string}) Path to qmd file to render.}

\item{rendered_assets}{(\code{character}) Paths to successfully rendered HTML
files to add to pkgdown menu.}

\item{root_dir}{(\code{string}) Path to the root directory of the package, used to
construct absolute paths. Default is \code{"."}.}

\item{source_dir}{(\code{string}) Path to the directory containing subdirectories
with \code{.*md} files to render. Default is \code{"pkgdown/menus"}.}

\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\description{
These parameters are used in multiple functions. They are defined here to
make them easier to import and to find.
}
\keyword{internal}
