% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-construct_menu.R
\name{empty_menu_metadata_table}
\alias{empty_menu_metadata_table}
\title{Construct an empty data frame of menu metadata}
\usage{
empty_menu_metadata_table()
}
\value{
A default blank data.frame with columns \code{html_file}, \code{title}, and
\code{index} for each md file in the menu subdirectory.
}
\description{
Construct an empty data frame of menu metadata
}
\keyword{internal}
