% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{ensure_pkgdown_examples_section}
\alias{ensure_pkgdown_examples_section}
\title{Ensure pkgdown examples menu exists}
\usage{
ensure_pkgdown_examples_section(pkgdown_yaml)
}
\arguments{
\item{pkgdown_yaml}{List representation of \verb{_pkgdown.yml}.}
}
\value{
Updated pkgdown YAML list.
}
\description{
Ensure pkgdown examples menu exists
}
\keyword{internal}
