% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{ensure_pkgdown_menu_section}
\alias{ensure_pkgdown_menu_section}
\title{Ensure pkgdown menu sections exist}
\usage{
ensure_pkgdown_menu_section(pkgdown_contents, menu)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{(\code{string}) Menu folder name.}
}
\value{
An updated list of pkgdown YAML contents with the new menu added.
}
\description{
Ensure pkgdown menu sections exist
}
\keyword{internal}
