% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{ensure_pkgdown_navbar_left}
\alias{ensure_pkgdown_navbar_left}
\title{Ensure pkgdown navbar structure left element exists}
\usage{
ensure_pkgdown_navbar_left(pkgdown_contents)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}
}
\value{
An updated list of pkgdown YAML contents with \code{structure$left} in
the navbar.
}
\description{
Ensure pkgdown navbar structure left element exists
}
\keyword{internal}
