% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-pkgdown.R
\name{extract_pattern}
\alias{extract_pattern}
\title{Turn a string into a regex pattern to match pieces of that string}
\usage{
extract_pattern(x)
}
\arguments{
\item{x}{A string in "camelCase", "snake_case", "kebab-case", "Title Case",
"Sentence case", or "lowercase" format. Strings of capital letters within
the string (e.g. "API" in "TheAPIClient") will be treated a group.}
}
\value{
A regex pattern to match pieces of the input string, ignoring
separators and case.
}
\description{
Turn a string into a regex pattern to match pieces of that string
}
\keyword{internal}
