% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-create_gha_manifest.R
\name{infer_from_yaml}
\alias{infer_from_yaml}
\title{Infer value from YAML/MD file
This helper function extracts a value from a YAML or Markdown file based on a specified type.}
\usage{
infer_from_yaml(path, type)
}
\arguments{
\item{path}{\code{string} path to the file}

\item{type}{\code{string} type of value to extract, as appears in file (e.g., "name", "type", "# Description")}
}
\value{
\code{string} extracted value or \code{NA_character_} if not found
}
\description{
Infer value from YAML/MD file
This helper function extracts a value from a YAML or Markdown file based on a specified type.
}
