% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{list_example_metadata}
\alias{list_example_metadata}
\title{List example metadata from Rmd files}
\usage{
list_example_metadata(rmd_dir)
}
\arguments{
\item{rmd_dir}{Character. Path to directory containing example Rmd files.}
}
\value{
Data frame with \code{html}, \code{title}, and \code{index} columns.
}
\description{
List example metadata from Rmd files
}
\keyword{internal}
