% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{list_non_index_html}
\alias{list_non_index_html}
\title{List HTML files excluding index.html}
\usage{
list_non_index_html(examples_dir, debug = FALSE)
}
\arguments{
\item{examples_dir}{Character. Path to directory containing example HTML files.}

\item{debug}{Logical. Whether to show debug messages.}
}
\value{
Character vector of HTML file names.
}
\description{
List HTML files excluding index.html
}
\keyword{internal}
