% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_examples.R
\name{normalize_output_dir}
\alias{normalize_output_dir}
\title{Normalize output directory}
\usage{
normalize_output_dir(output_dir)
}
\arguments{
\item{output_dir}{Character. Path to output directory.}
}
\value{
Absolute path to output directory.
}
\description{
Normalize output directory
}
\keyword{internal}
