% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example.R
\name{or_pipe}
\alias{or_pipe}
\alias{\%||\%}
\title{Helper for missing values}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to test.}

\item{y}{Fallback value.}
}
\value{
\code{x} if not \code{NULL}, otherwise \code{y}.
}
\description{
Helper for missing values
}
\keyword{internal}
