% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_assets.R
\name{render_assets}
\alias{render_assets}
\title{Render Rmd and qmd files in a menu subdirectory}
\usage{
render_assets(menu_subdir, assets_dir, verbose)
}
\arguments{
\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{assets_dir}{(\code{string}) Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
Character vector of relative paths to rendered HTML files for
successfully rendered Rmd and qmd files.
}
\description{
Render Rmd and qmd files in a menu subdirectory
}
\keyword{internal}
