% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_qmd.R
\name{render_qmd}
\alias{render_qmd}
\title{Render an individual qmd file}
\usage{
render_qmd(qmd_file, output_file, params = NULL, verbose = FALSE)
}
\arguments{
\item{qmd_file}{Character. Path to qmd file to render.}

\item{output_file}{Character. Path to output HTML file to create.}

\item{params}{List. Optional list of parameters to pass to
\code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\value{
\code{output_file} (on success) or \code{""} (on failure), invisibly.
}
\description{
Render an individual qmd file
}
\keyword{internal}
