% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-check_workflow_compliance.R
\name{report_compliance_results}
\alias{report_compliance_results}
\title{Report Compliance Results}
\usage{
report_compliance_results(
  missing_workflows,
  extra_workflows,
  version_issues,
  content_issues,
  has_errors
)
}
\arguments{
\item{missing_workflows}{\code{character} vector of missing workflows}

\item{extra_workflows}{\code{character} vector of extra workflows}

\item{version_issues}{\code{character} vector of version issues}

\item{content_issues}{\code{character} vector of content issues}

\item{has_errors}{\code{logical} whether critical errors were found}
}
\value{
NULL (prints results to console)
}
\description{
Internal function to report workflow compliance check results.
}
\keyword{internal}
