% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-util.R
\name{to_title_case}
\alias{to_title_case}
\title{Convert a string to title case, preserving acronyms}
\usage{
to_title_case(x)
}
\arguments{
\item{x}{A string in "camelCase", "snake_case", "kebab-case", "Title Case",
"Sentence case", or "lowercase" format. Strings of capital letters within
the string (e.g. "API" in "TheAPIClient") will be treated a group.}
}
\value{
The string in Title Case, with separators removed and acronyms
preserved. For example, "theAPIClient" and "the_API_client" both become
"The API Client".
}
\description{
Convert a string to title case, preserving acronyms
}
\keyword{internal}
