% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{update_pkgdown_examples}
\alias{update_pkgdown_examples}
\title{Update pkgdown YAML with example menu entries}
\usage{
update_pkgdown_examples(pkgdown_yml, html_files, metadata)
}
\arguments{
\item{pkgdown_yml}{Character. Path to \verb{_pkgdown.yml}.}

\item{html_files}{Character vector of example HTML files.}

\item{metadata}{Data frame of example metadata.}
}
\value{
\code{NULL} invisibly.
}
\description{
Update pkgdown YAML with example menu entries
}
\keyword{internal}
