% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{add_action}
\alias{add_action}
\title{Add a Gilead GitHub Action to package}
\usage{
add_action(
  name,
  version,
  ...,
  workflows_path = "./.github/workflows",
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{name}{String. The action to install.}

\item{version}{String. The expected version of the action.}

\item{...}{These dots are for future extensions and must be empty.}

\item{workflows_path}{String. Path to the package workflows.}

\item{overwrite}{Logical. Overwrite existing files?}

\item{verbose}{Logical. Inform about changes?}
}
\value{
The name of the workflow if it was updated, otherwise an empty
character vector.
}
\description{
Add an official Gilead GitHub Action from
\url{https://github.com/Gilead-BioStats/gsm.utils@actions-v1} to a package, or
update an existing Gilead GitHub Actions to the latest version if necessary.
}
