% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{add_actions}
\alias{add_actions}
\title{Add Gilead GitHub Actions to package}
\usage{
add_actions(strPackageDir = ".", overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{overwrite}{(\code{boolean}) Overwrite existing files?}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
A character vector of added and updated action names, invisibly.
}
\description{
Add the official Gilead GitHub Actions to a package, and update existing
Gilead GitHub Actions to the latest versions if necessary.
}
