% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{add_contributor_guidelines}
\alias{add_contributor_guidelines}
\title{Add GSM Contributor Guidelines markdown to package}
\usage{
add_contributor_guidelines(strPackageDir = ".", overwrite = TRUE)
}
\arguments{
\item{strPackageDir}{path to package directory}

\item{overwrite}{\code{boolean} argument declaring whether or not to overwrite existing files. Default is \code{TRUE}.}
}
\description{
Add GSM Contributor Guidelines markdown to package
}
