% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{add_gsm_actions}
\alias{add_gsm_actions}
\title{Add GSM GitHub Actions to package}
\usage{
add_gsm_actions(strPackageDir = ".", overwrite = TRUE)
}
\arguments{
\item{strPackageDir}{path to package directory}

\item{overwrite}{\code{boolean} argument declaring whether or not to overwrite
existing files. Default is \code{TRUE}.}
}
\description{
Add GSM GitHub Actions to package
}
