% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{add_gsm_issue_templates}
\alias{add_gsm_issue_templates}
\title{Add GSM issue templates to package}
\usage{
add_gsm_issue_templates(strPackageDir = ".", overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{overwrite}{(\code{boolean}) Overwrite existing files?}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\description{
Add GSM issue templates to package
}
