% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-build_assets.R
\name{add_menu}
\alias{add_menu}
\title{Add a menu to pkgdown from a subdirectory of assets}
\usage{
add_menu(menu_subdir, assets_dir, pkgdown_yml, verbose)
}
\arguments{
\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{assets_dir}{(\code{string}) Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}

\item{pkgdown_yml}{(\code{string}) Path to the \verb{_pkgdown.yml} file to update with
new menu items, or from which to remove unused menu items.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
\code{NULL} (invisibly). Called for the side effect of updating
\code{pkgdown_yml} with new menu items based on rendered assets.
}
\description{
Add a menu to pkgdown from a subdirectory of assets
}
\keyword{internal}
