% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{add_pkgdown_examples}
\alias{add_pkgdown_examples}
\title{Generate Examples menu in pkgdown}
\usage{
add_pkgdown_examples(
  examples_dir = "pkgdown/assets/examples",
  pkgdown_yml = "_pkgdown.yml"
)
}
\arguments{
\item{examples_dir}{Character. Path to directory containing example HTML files.
Default is "pkgdown/assets/examples".}

\item{pkgdown_yml}{Character. Path to _pkgdown.yml file to update with menu.
Default is "_pkgdown.yml".}
}
\value{
Invisibly returns NULL.
}
\description{
Scans for HTML files in a directory (default is pkgdown/assets/examples) and updates
a standard _pkgdown.yml with a new sub-menu listing the html files.
}
