% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{add_pkgdown_examples}
\alias{add_pkgdown_examples}
\title{Generate Examples menu in pkgdown}
\usage{
add_pkgdown_examples(
  examples_dir = "pkgdown/assets/examples",
  pkgdown_yml = "_pkgdown.yml",
  rmd_dir = "inst/examples"
)
}
\arguments{
\item{examples_dir}{Character. Path to directory containing example HTML
files. Default is \code{"pkgdown/assets/examples"}.}

\item{pkgdown_yml}{Character. Path to \verb{_pkgdown.yml} file to update with menu.
Default is \code{"_pkgdown.yml"}.}

\item{rmd_dir}{Character. Path to directory containing example \verb{\\.Rmd} files
used to derive titles and order. Default is \code{"inst/examples"}.}
}
\value{
\code{NULL} invisibly.
}
\description{
Scans for HTML files in a directory (default is \code{pkgdown/assets/examples})
and updates a standard \verb{_pkgdown.yml} with a new sub-menu listing the html
files.
}
