% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_assets.R
\name{build_assets}
\alias{build_assets}
\title{Build additional pkgdown assets}
\usage{
build_assets(
  source_dir = "pkgdown/menus",
  assets_dir = "pkgdown/assets",
  pkgdown_yml = "_pkgdown.yml",
  root_dir = ".",
  verbose = FALSE
)
}
\arguments{
\item{source_dir}{Character. Path to the directory containing subdirectories
with \code{.*md} files to render. Default is \code{"pkgdown/menus"}.}

\item{assets_dir}{Character. Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}

\item{pkgdown_yml}{Character. Path to the \verb{_pkgdown.yml} file to update with
new menu items, or from which to remove unused menu items.}

\item{root_dir}{Character. Path to the root directory of the package, used to
construct absolute paths. Default is \code{"."}.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\value{
\code{NULL} (invisibily). Called for the side effect of setting up
pkgdown assets.
}
\description{
Scan a directory for subdirectories, render any \code{.*md} files in those
subdirectories to \code{.html}, and add them to the pkgdown index.
}
