% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-build_assets.R
\name{build_assets_params}
\alias{build_assets_params}
\title{Asset params}
\arguments{
\item{assets_dir}{Character. Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}

\item{menu}{Character. Menu folder name.}

\item{menu_subdir}{Character. Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{output_file}{Character. Path to output HTML file to create.}

\item{params}{List. Optional list of parameters to pass to
\code{\link[quarto:quarto_render]{quarto::quarto_render()}}.}

\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{pkgdown_yml}{Character. Path to the \verb{_pkgdown.yml} file to update with
new menu items, or from which to remove unused menu items.}

\item{qmd_file}{Character. Path to qmd file to render.}

\item{rendered_assets}{Character. Path to successfully rendered HTML files to
add to pkgdown menu.}

\item{root_dir}{Character. Path to the root directory of the package, used to
construct absolute paths. Default is \code{"."}.}

\item{source_dir}{Character. Path to the directory containing subdirectories
with \code{.*md} files to render. Default is \code{"pkgdown/menus"}.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\description{
Asset params
}
\keyword{internal}
