% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example.R
\name{build_example_filename}
\alias{build_example_filename}
\title{Build a safe example filename}
\usage{
build_example_filename(strName, strType)
}
\arguments{
\item{strName}{Character. Display name of the example.}

\item{strType}{Character. Example type prefix.}
}
\value{
File name for the example.
}
\description{
Build a safe example filename
}
\keyword{internal}
