% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-create_gha_manifest.R
\name{create_gha_manifest}
\alias{create_gha_manifest}
\title{Create GitHub Actions Manifest
This function generates a JSON manifest file summarizing the GitHub Actions workflows}
\usage{
create_gha_manifest(
  workflows_dir = "inst/gha_templates/workflows",
  issue_templates_dir = "inst/gha_templates/ISSUE_TEMPLATE",
  repository = "https://github.com/Gilead-BioStats/gsm.utils",
  description_path = "DESCRIPTION",
  output_path = "inst/gha_templates/gha_version.json"
)
}
\arguments{
\item{workflows_dir}{\code{string} path to directory containing GitHub Actions workflow \code{.yaml} files}

\item{issue_templates_dir}{\code{string} path to directory containing GitHub issue template \code{.md} files}

\item{repository}{\code{string} URL of the GitHub repository}

\item{description_path}{\code{string} path to \code{DESCRIPTION} file}

\item{output_path}{\code{string} path to output JSON manifest file}
}
\description{
Create GitHub Actions Manifest
This function generates a JSON manifest file summarizing the GitHub Actions workflows
}
