% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{ensure_pkgdown_components_menu}
\alias{ensure_pkgdown_components_menu}
\title{Ensure pkgdown components element has menu}
\usage{
ensure_pkgdown_components_menu(pkgdown_contents, menu)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{Character. Menu folder name.}
}
\value{
An updated list of pkgdown YAML contents with the menu in navbar
components.
}
\description{
Ensure pkgdown components element has menu
}
\keyword{internal}
