% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{ensure_pkgdown_navbar_left_menu}
\alias{ensure_pkgdown_navbar_left_menu}
\title{Ensure pkgdown navbar structure left element contains menu}
\usage{
ensure_pkgdown_navbar_left_menu(pkgdown_contents, menu)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{Character. Menu folder name.}
}
\value{
An updated list of pkgdown YAML contents with the menu in
\code{navbar$structure$left}.
}
\description{
Ensure pkgdown navbar structure left element contains menu
}
\keyword{internal}
