% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{filter_existing_menu}
\alias{filter_existing_menu}
\title{Filter existing menu items to only those with assets that still exist}
\usage{
filter_existing_menu(pkgdown_contents, menu, assets_dir)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{(\code{string}) Menu folder name.}

\item{assets_dir}{(\code{string}) Path to the directory where rendered \code{.html}
files should be saved. Default is \code{"pkgdown/assets"}.}
}
\value{
A list of existing menu items from the pkgdown YAML contents that
correspond to assets that still exist in the assets directory.
}
\description{
Filter existing menu items to only those with assets that still exist
}
\keyword{internal}
