% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.utils-package.R
\docType{package}
\name{gsm.utils-package}
\alias{gsm.utils}
\alias{gsm.utils-package}
\title{gsm.utils: Utilities for gsm package developers}
\description{
Provides utility functions and file templates to be used for the set-up and development of gsm packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gilead-biostats.github.io/gsm.utils}
  \item \url{https://github.com/Gilead-BioStats/gsm.utils}
  \item Report bugs at \url{https://github.com/Gilead-BioStats/gsm.utils/issues}
}

}
\author{
\strong{Maintainer}: Jeremy Wildfire \email{jwildfire@gmail.com}

Authors:
\itemize{
  \item Jeremy Wildfire \email{jwildfire@gmail.com}
  \item Jon Harmon \email{jon.harmon@atorusresearch.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})
  \item Laura Maxwell \email{lkmaxwell23@gmail.com}
  \item Natalia Andriychuk \email{natalia.andriychuk@gilead.com}
  \item Anne Zheng \email{anne.zheng4@gilead.com}
  \item Spencer Childress \email{Spencer.Childress@gilead.com}
}

Other contributors:
\itemize{
  \item Gilead Sciences [copyright holder]
}

}
\keyword{internal}
