% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsm.utils-package.R
\docType{package}
\name{gsm.utils-package}
\alias{gsm.utils}
\alias{gsm.utils-package}
\title{gsm.utils: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
