% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-init_gsm_package.R
\name{init_gsm_package}
\alias{init_gsm_package}
\title{Initialize gsm Extension package}
\usage{
init_gsm_package(
  strPackageDir,
  lDescriptionFields = list(),
  bIncludeWorkflowDir = TRUE,
  strOrg = "Gilead-Public"
)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{lDescriptionFields}{(\code{list}) Description fields, passed to
\code{\link[usethis:create_package]{usethis::create_package()}}. Default is \code{list()}.}

\item{bIncludeWorkflowDir}{(\code{boolean}) Whether or not to include the
\code{inst/workflow} directory in the root of the package. Default is \code{TRUE}.}

\item{strOrg}{(\code{string}) GitHub organization under which the repo should be
created. Set to \code{NULL} to create the package in your personal GitHub
account.}
}
\description{
Initialize gsm Extension package
}
