% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-init_gsm_package.R
\name{init_gsm_package}
\alias{init_gsm_package}
\title{Initialize gsm Extension package}
\usage{
init_gsm_package(
  strPackageDir,
  lDescriptionFields = list(),
  bIncludeWorkflowDir = TRUE
)
}
\arguments{
\item{strPackageDir}{path to package directory}

\item{lDescriptionFields}{\code{list} of description fields, passed to \code{\link[usethis:create_package]{usethis::create_package()}}. Default is \code{list()}.}

\item{bIncludeWorkflowDir}{\code{boolean} argument declaring whether or not to include the \code{inst/workflow} directory in the root of the package. Default is \code{TRUE}.}
}
\description{
Initialize gsm Extension package
}
