% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_example.R
\name{make_example}
\alias{make_example}
\title{Create an example R Markdown template}
\usage{
make_example(
  strName = "Example_Name",
  strType = c("Example", "Cookbook"),
  strDetails = "<<Fill in Example description here>>",
  intIndex = 999,
  output_dir = "pkgdown/menus/examples",
  overwrite = FALSE
)
}
\arguments{
\item{strName}{(\code{string}) Display name of the example.}

\item{strType}{(\code{string}) Type of example, either \code{"Example"} or
\code{"Cookbook"}.}

\item{strDetails}{(\code{string}) Optional description for the example.}

\item{intIndex}{(\code{numeric}) Optional ordering index for the examples menu.}

\item{output_dir}{(\code{string}) Directory to write the example to.}

\item{overwrite}{(\code{boolean}) Overwrite existing files? Default is \code{TRUE}.}
}
\value{
Path to the created example file (invisibly).
}
\description{
Creates a new example \code{.Rmd} file from a standard template.
}
