% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_slides.R
\name{make_slides}
\alias{make_slides}
\title{Create a slide deck template}
\usage{
make_slides(
  strTitle = "Slide Deck Title",
  intIndex = 999,
  output_dir = "pkgdown/menus/slides",
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{strTitle}{(\code{string}) Display title.}

\item{intIndex}{(\code{numeric}) Sort-order index. \code{NULL}, \code{NA}, and length-0
vectors are treated as absent.}

\item{output_dir}{(\code{string}) Directory to write the asset to.}

\item{overwrite}{(\code{boolean}) Overwrite existing files?}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
Path to the created slide deck file (invisibly).
}
\description{
Creates a new RevealJS \code{.qmd} file from a standard template.
}
