% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{remove_deprecated_actions}
\alias{remove_deprecated_actions}
\title{Remove deprecated Gilead GitHub Actions from package}
\usage{
remove_deprecated_actions(
  strPackageDir = ".",
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{strPackageDir}{(\code{string}) Path to the package directory.}

\item{overwrite}{(\code{boolean}) Overwrite existing files? Default is \code{TRUE}.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
A character vector of deleted action names, invisibly.
}
\description{
Remove old workflows that we no longer recommend nor support:
"R-CMD-check-dev.yaml", "pkgdown-cleanup.yaml", "pkgdown-with-examples.yaml",
"r-releaser.yaml", and "r_releaser.yaml".
}
