% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.R
\name{remove_deprecated_actions}
\alias{remove_deprecated_actions}
\title{Remove deprecated Gilead GitHub Actions from package}
\usage{
remove_deprecated_actions(
  strPackageDir = ".",
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{strPackageDir}{String. Path to package directory}

\item{overwrite}{Logical. Is it ok to delete existing files?}

\item{verbose}{Logical. Inform about changes?}
}
\value{
A character vector of deleted action names, invisibly.
}
\description{
Remove old workflows that we no longer recommend nor support:
"R-CMD-check-dev.yaml", "pkgdown-cleanup.yaml", "pkgdown-with-examples.yaml",
"r-releaser.yaml", and "r_releaser.yaml".
}
