% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{remove_pkgdown_examples}
\alias{remove_pkgdown_examples}
\title{Remove examples menu from pkgdown YAML}
\usage{
remove_pkgdown_examples(pkgdown_yml, examples_dir)
}
\arguments{
\item{pkgdown_yml}{Character. Path to \verb{_pkgdown.yml}.}

\item{examples_dir}{Character. Path to examples directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Remove examples menu from pkgdown YAML
}
\keyword{internal}
