% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{remove_pkgdown_menu}
\alias{remove_pkgdown_menu}
\title{Remove menu from pkgdown YAML}
\usage{
remove_pkgdown_menu(pkgdown_contents, menu, verbose)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{(\code{string}) Menu folder name.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
An updated list of pkgdown YAML contents with the menu removed.
}
\description{
Remove menu from pkgdown YAML
}
\keyword{internal}
