% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_qmd.R
\name{render_qmd_assets}
\alias{render_qmd_assets}
\title{Render qmd files in a menu subdirectory}
\usage{
render_qmd_assets(menu_subdir, output_dir, verbose)
}
\arguments{
\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{output_dir}{(\code{string}) Directory to write the example to.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
Character vector of paths to rendered HTML files for successfully
rendered qmd files, or \code{""} for failed renders.
}
\description{
Render qmd files in a menu subdirectory
}
\keyword{internal}
