% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_qmd.R
\name{render_qmd_assets}
\alias{render_qmd_assets}
\title{Render qmd files in a menu subdirectory}
\usage{
render_qmd_assets(menu_subdir, output_dir, verbose)
}
\arguments{
\item{menu_subdir}{Character. Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{output_dir}{Character. Directory to write the example to.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\value{
Character vector of relative paths to rendered HTML files for
successfully rendered qmd files.
}
\description{
Render qmd files in a menu subdirectory
}
\keyword{internal}
