% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_rmd}
\alias{render_rmd}
\title{Custom Rmarkdown render function}
\usage{
render_rmd(
  strInputPath,
  strOutputFile = basename(strInputPath),
  strOutputDir = getwd(),
  lParams = NULL,
  quiet = FALSE
)
}
\arguments{
\item{strInputPath}{\code{string} or \code{fs_path} Path to the template \code{Rmd} file.}

\item{strOutputFile}{\code{string} Filename for the output.}

\item{strOutputDir}{\code{string} or \code{fs_path} Path to the directory where the
output will be saved.}

\item{lParams}{\code{list} Parameters to pass to the template \code{Rmd} file.}

\item{quiet}{Logical. Passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Rendered Rmarkdown file path (invisibly).
}
\description{
Rmarkdown render function that defaults to rendering intermediate Rmd files
in a temporary directory, and falls back to a writable output directory when
needed.
}
