% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_rmd_assets}
\alias{render_rmd_assets}
\title{Render Rmd files in a menu subdirectory}
\usage{
render_rmd_assets(menu_subdir, output_dir, verbose)
}
\arguments{
\item{menu_subdir}{(\code{string}) Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{output_dir}{(\code{string}) Directory to write the example to.}

\item{verbose}{(\code{boolean}) Inform about changes? Default is \code{TRUE}.}
}
\value{
Character vector of relative paths to rendered HTML files for
successfully rendered Rmd files.
}
\description{
Render Rmd files in a menu subdirectory
}
\keyword{internal}
