% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-util.R
\name{resolve_md_title}
\alias{resolve_md_title}
\title{Determine a title for a file}
\usage{
resolve_md_title(path, title = NULL)
}
\arguments{
\item{path}{Character. The path to a file.}

\item{title}{Character. An optional string to use as the title.}
}
\value{
A title-case string to use as the title of the file, without html,
etc.
}
\description{
Determine a title for a file
}
\keyword{internal}
