% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{update_gsm_package}
\alias{update_gsm_package}
\title{Update GSM package with global issue templates and GH actions}
\usage{
update_gsm_package(strPackageDir = ".", overwrite = TRUE, verbose = TRUE)
}
\arguments{
\item{strPackageDir}{path to package directory}

\item{overwrite}{\code{boolean} argument declaring whether or not to overwrite
existing files. Default is \code{TRUE}.}

\item{verbose}{\code{boolean} argument declaring whether to emit messages about
updates.}
}
\description{
Update GSM package with global issue templates and GH actions
}
