% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_gsm_package.R
\name{update_gsm_package}
\alias{update_gsm_package}
\title{Update GSM package with global issue templates and GH actions}
\usage{
update_gsm_package(strPackageDir = ".")
}
\arguments{
\item{strPackageDir}{path to package directory}
}
\description{
Update GSM package with global issue templates and GH actions
}
