% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown-add_pkgdown_nav.R
\name{update_pkgdown_menu}
\alias{update_pkgdown_menu}
\title{Update pkgdown contents with menu}
\usage{
update_pkgdown_menu(
  pkgdown_contents,
  menu,
  menu_subdir,
  rendered_assets,
  existing_menu,
  verbose
)
}
\arguments{
\item{pkgdown_contents}{List. The contents of \verb{_pkgdown.yml}, as loaded by
\code{\link[yaml:read_yaml]{yaml::read_yaml()}}.}

\item{menu}{Character. Menu folder name.}

\item{menu_subdir}{Character. Path to the subdirectory containing \code{.*md}
files to render and add to pkgdown.}

\item{rendered_assets}{Character. Path to successfully rendered HTML files to
add to pkgdown menu.}

\item{existing_menu}{List. Existing menu items from the pkgdown YAML
contents, if any, to preserve when adding new menu items for rendered
assets.}

\item{verbose}{Logical. Whether to print messages about rendered assets and
menu updates.}
}
\value{
An updated list of pkgdown YAML contents with the new menu entries
added.
}
\description{
Update pkgdown contents with menu
}
\keyword{internal}
