% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-update_workflow_version.R
\name{update_workflow_version}
\alias{update_workflow_version}
\title{Update version in GitHub Actions workflow \code{.yaml} files}
\usage{
update_workflow_version(
  dir = "inst/gha_templates/workflows",
  description_path = "DESCRIPTION",
  label = "# gsm.utils GHA version",
  recursive = TRUE,
  add_if_missing = TRUE
)
}
\arguments{
\item{dir}{\code{string} path to directory containing GitHub Actions workflow \code{.yaml} files}

\item{description_path}{\code{string} path to \code{DESCRIPTION} file}

\item{label}{\code{string} label to identify version line in workflow files}

\item{recursive}{\code{boolean} whether to search directories recursively}

\item{add_if_missing}{\code{boolean} whether to add version line if missing}
}
\description{
Update version in GitHub Actions workflow \code{.yaml} files
}
