% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pkgdown_examples.R
\name{write_yaml}
\alias{write_yaml}
\title{Helper to wrap yaml::write_yaml for testing.}
\usage{
write_yaml(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[yaml:write_yaml]{yaml::write_yaml()}}.}
}
\value{
\code{NULL} (invisibly)
}
\description{
Helper to wrap yaml::write_yaml for testing.
}
\keyword{internal}
